requirements.txt
# Dépendances pour l'application Photos Manager

# Traitement d'images
Pillow>=10.0.0
piexif>=1.1.3

# Traitement GPX et cartographie
gpxpy>=1.5.0
matplotlib>=3.7.0
contextily>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0

# Requêtes HTTP
requests>=2.31.0

# Calculs numériques
numpy>=1.24.0

# Support des fuseaux horaires (intégré à Python 3.9+)
# Si Python < 3.9, décommenter la ligne suivante :
# backports.zoneinfo>=0.2.1
